Improve exists query rewrite.
The ExistsQueryBuilder rewrite only depends on information that exists in mappings. Therefor the rewrite can be done as index metadata rewrite. Practically this means the rewrite can occur during can match phase before opening a searcher or doing any potential refresh.

The `getMatchingFieldNames(...)` method was moved from SearchExecutionContext to QueryRewriteContext.

As an additional change, the QueryRewriteContext#doSearchRewrite(...) method now by default delegates to doIndexMetadataRewrite(...). This is because a search rewrite by default should be compatible with a index metadata rewrite.
